+++ nutree/fs.py
@@ -32,7 +32,7 @@ class FileSystemEntry:
         self.size = int(size)
         self.mdate = float(mdate) if mdate is not None else None
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         if self.is_dir:
             return f"[{self.name}]"
         assert self.mdate is not None
@@ -44,7 +44,7 @@ class FileSystemTree(Tree[FileSystemEntry]):
     DEFAULT_KEY_MAP = {}  # don't replace 's' with 'str'
 
     @classmethod
-    def serialize_mapper(cls, node: Node, data: dict):
+    def serialize_mapper(cls, node: Node, data: dict) -> dict:
         """Callback for use with :meth:`~nutree.tree.Tree.save`."""
         inst = node.data
         if inst.is_dir:
@@ -54,7 +54,7 @@ class FileSystemTree(Tree[FileSystemEntry]):
         return data
 
     @classmethod
-    def deserialize_mapper(cls, parent: Node, data: dict):
+    def deserialize_mapper(cls, parent: Node, data: dict) -> FileSystemEntry:
         """Callback for use with :meth:`~nutree.tree.Tree.load`."""
         # v = data["v"]
         if "d" in data:
@@ -72,7 +72,7 @@ def load_tree_from_fs(path: str | Path, *, sort: bool = True) -> FileSystemTree:
     path = Path(path)
     tree = FileSystemTree(str(path))
 
-    def visit(node: Node, pth: Path):
+    def visit(node: Node, pth: Path) -> None:
         if sort:
             dirs = []
             files = []