+++ tests/test_bench.py
@@ -5,7 +5,6 @@
 # ruff: noqa: E501 Line too long
 
 import sys
-from pathlib import Path
 
 import pytest
 from nutree import IterMethod
@@ -20,147 +19,113 @@ benchmark = pytest.mark.skipif(
 
 @benchmark
 class TestBenchmarks:
-    def test_bench_index(self, capsys):
+    def test_bench_index(self, benchman, capsys):
         """ """
-        results = ["Benchmark results"]
-        tree = fixture.create_tree_simple()
-
-        results.append(
-            fixture.run_timings(
-                "access tree[key]",
-                """\
-            _ = tree["a1"]
-        """,
-                globals=locals(),
-                time_unit="nsec",
-            )
+        tree = fixture.create_tree_simple(print=False)
+        bmr = benchman.make_runner(
+            name="search",
+            sample_size=len(tree),
+            globals=locals(),
         )
 
-        results.append(
-            fixture.run_timings(
-                "tree.find()",
-                """\
-            _ = tree.find("a1")
-        """,
-                globals=locals(),
-            )
+        bmr.run(
+            variant="by index",
+            stmt="""\
+                _ = tree["a1"]
+            """,
         )
-
-        results.append(
-            fixture.run_timings(
-                "tree.find_all()",
-                """\
-            _ = tree.find_all("a1")
-        """,
-                globals=locals(),
-            )
+        bmr.run(
+            variant="find() ",
+            stmt="""\
+                _ = tree.find("a1")
+            """,
+        )
+        bmr.run(
+            variant="find_all() ",
+            stmt="""\
+                _ = tree.find_all("a1")
+            """,
         )
 
         with capsys.disabled():
-            print("\n  - ".join(results))
+            bmr.print()
 
-    def test_bench_properties(self, capsys):
+    def test_bench_properties(self, benchman, capsys):
         """ """
-
-        results = ["Benchmark results"]
         tree = fixture.create_tree_simple()
         node = tree.first_child()
-
-        results.append(
-            fixture.run_timings(
-                "access node.data (property)",
-                """\
-            _ = node.data
-        """,
-                globals=locals(),
-            )
+        bmr = benchman.make_runner(
+            name="access node.data",
+            globals=locals(),
         )
 
-        results.append(
-            fixture.run_timings(
-                "access node._data (attribute)",
-                """\
-            _ = node._data
-        """,
-                globals=locals(),
-            )
+        bmr.run(
+            variant="node.data (property)",
+            stmt="""\
+                _ = node.data
+            """,
+        )
+
+        bmr.run(
+            variant="node._data (attr)",
+            stmt="""\
+                _ = node._data
+            """,
         )
 
         with capsys.disabled():
-            print("\n  - ".join(results))
+            bmr.print()
 
-    def test_bench_iter(self, capsys):
+    def test_bench_iter(self, benchman, capsys):
         """ """
-
-        results = ["Benchmark results"]
         tree = fixture.create_tree_simple()
         node = tree.first_child()
-
-        results.append(
-            fixture.run_timings(
-                "for _ in tree: ...",
-                """\
-            for _ in tree: pass
-        """,
-                globals=locals(),
-            )
+        size = len(tree)
+        bmr = benchman.make_runner(
+            name="iterate",
+            sample_size=len(tree),
+            globals=locals(),
         )
 
-        results.append(
-            fixture.run_timings(
-                "for _ in tree.iterator(): ...",
-                """\
-            for _ in tree.iterator(): pass
-        """,
-                globals=locals(),
-            )
+        bmr.run(
+            variant="for _ in tree: ...",
+            stmt="""\
+                for _ in tree: pass
+            """,
         )
 
-        ORDER = IterMethod.LEVEL_ORDER
-        results.append(
-            fixture.run_timings(
-                f"for _ in tree.iterator({ORDER}): ...",
-                """\
-            for _ in tree.iterator(ORDER): pass
-        """,
-                globals=locals(),
-            )
+        bmr.run(
+            variant="for _ in tree.iterator(): ...",
+            stmt="""\
+                for _ in tree.iterator(): pass
+            """,
         )
 
-        ORDER = IterMethod.RANDOM_ORDER
-        results.append(
-            fixture.run_timings(
-                f"for _ in tree.iterator({ORDER}): ...",
-                """\
-            for _ in tree.iterator(ORDER): pass
-        """,
+        ORDER_LIST = [
+            IterMethod.LEVEL_ORDER,
+            IterMethod.PRE_ORDER,
+            IterMethod.POST_ORDER,
+            IterMethod.RANDOM_ORDER,
+            IterMethod.UNORDERED,
+        ]
+        for ORDER in ORDER_LIST:
+            bmr.run(
+                variant=f"for _ in tree.iterator({ORDER.name}): ...",
+                stmt="""\
+                    for _ in tree.iterator(ORDER): pass
+                """,
                 globals=locals(),
             )
-        )
 
-        ORDER = IterMethod.UNORDERED
-        results.append(
-            fixture.run_timings(
-                f"for _ in tree.iterator({ORDER}): ...",
-                """\
-            for _ in tree.iterator(ORDER): pass
-        """,
-                globals=locals(),
-            )
-        )
-
-        results.append(
-            fixture.run_timings(
-                "tree.visit(lambda node, memo: None)",
-                """\
-            tree.visit(lambda node, memo: None)
-        """,
-                globals=locals(),
-            )
+        bmr.run(
+            variant="tree.visit(lambda node, memo: None)",
+            stmt="""\
+                tree.visit(lambda node, memo: None)
+            """,
         )
 
         with capsys.disabled():
-            print("\n  - ".join(results))
+            bmr.print()
 
 
 @benchmark
@@ -176,118 +141,100 @@ class TestCompress:
     ZIP_LZMA ('.lzma')     |       44.347 |       2.13 |    0.38
     """
 
-    def test_bench_serialize(self, capsys):
+    def test_bench_serialize(self, benchman, tmp_path, capsys):
         import zipfile  # noqa: F401
 
-        results = ["Benchmark results"]
-        directory = Path("~").expanduser()
-        # with WritableTempFile("w", suffix=".gz") as temp_file:
+        # directory = Path("~").expanduser()
+        directory = tmp_path
 
         tree = fixture.generate_tree([10, 100, 100])
 
+        bmr_load = benchman.make_runner(
+            name="serialize_load",
+            # sample_size=len(tree),
+            repeat=1,
+            iterations=1,
+            globals=locals(),
+        )
+        bmr_save = benchman.make_runner(
+            name="serialize_save",
+            # sample_size=len(tree),
+            repeat=1,
+            iterations=1,
+            globals=locals(),
+        )
+
         path = directory / "test.json"
-        results.append(
-            fixture.run_timings(
-                f"tree.save({path}, compression=False, nodes={len(tree):,})",
-                f"""\
+        bmr_save.run(
+            variant="uncompressed ('.json')",
+            stmt=f"""\
                 tree.save("{path}", compression=False)
-                """,
-                repeat=1,
-                number=1,
-                globals=locals(),
-            )
+            """,
         )
-        results.append(
-            fixture.run_timings(
-                f"tree.load({path}, nodes={len(tree):,})",
-                f"""\
+
+        bmr_load.run(
+            variant="uncompressed ('.json')",
+            stmt=f"""\
                 tree.load("{path}")
-                """,
-                repeat=1,
-                number=1,
-                globals=locals(),
-            )
+            """,
         )
+
         # --- ZIP
         path = directory / "test.zip"
-        results.append(
-            fixture.run_timings(
-                f"tree.save({path}, compression=ZIP, nodes={len(tree):,})",
-                f"""\
+        bmr_save.run(
+            variant="ZIP_DEFLATED ('.zip') ",
+            stmt=f"""\
                 tree.save("{path}", compression=zipfile.ZIP_DEFLATED)
-                """,
-                repeat=1,
-                number=1,
-                globals=locals(),
-            )
+            """,
         )
-        results.append(
-            fixture.run_timings(
-                f"tree.load({path}, nodes={len(tree):,})",
-                f"""\
+
+        bmr_load.run(
+            variant="ZIP_DEFLATED ('.zip') ",
+            stmt=f"""\
                 tree.load("{path}")
-                """,
-                repeat=1,
-                number=1,
-                globals=locals(),
-            )
+            """,
         )
+
         # --- ZIP
         path = directory / "test.bz2"
-        results.append(
-            fixture.run_timings(
-                f"tree.save({path}, compression=ZIP, nodes={len(tree):,})",
-                f"""\
+        bmr_save.run(
+            variant="ZIP_BZIP2 ('.bz2') ",
+            stmt=f"""\
                 tree.save("{path}", compression=zipfile.ZIP_BZIP2)
-                """,
-                repeat=1,
-                number=1,
-                globals=locals(),
-            )
+            """,
         )
-        results.append(
-            fixture.run_timings(
-                f"tree.load({path}, nodes={len(tree):,})",
-                f"""\
+
+        bmr_load.run(
+            variant="ZIP_BZIP2 ('.bz2') ",
+            stmt=f"""\
                 tree.load("{path}")
-                """,
-                repeat=1,
-                number=1,
-                globals=locals(),
-            )
+            """,
         )
+
         # --- ZIP
         path = directory / "test.lzma"
-        results.append(
-            fixture.run_timings(
-                f"tree.save({path}, compression=LZMA, nodes={len(tree):,})",
-                f"""\
+        bmr_save.run(
+            variant="ZIP_LZMA ('.lzma') ",
+            stmt=f"""\
                 tree.save("{path}", compression=zipfile.ZIP_LZMA)
-                """,
-                repeat=1,
-                number=1,
-                globals=locals(),
-            )
+            """,
         )
-        results.append(
-            fixture.run_timings(
-                f"tree.load({path}, nodes={len(tree):,})",
-                f"""\
+
+        bmr_load.run(
+            variant="ZIP_LZMA ('.lzma') ",
+            stmt=f"""\
                 tree.load("{path}")
-                """,
-                repeat=1,
-                number=1,
-                globals=locals(),
-            )
+            """,
         )
 
         with capsys.disabled():
-            print("\n  - ".join(results))
+            bmr_save.print()
+            bmr_load.print()
 
 
 @benchmark
 class TestMemory:
-    def test_bench_node_size(self, capsys):
+    def test_bench_node_size(self, benchman, capsys):
         """ """
 
         try:
@@ -323,7 +270,8 @@ class TestMemory:
             f"node-size: {node_size:.1f} bytes"
         )
         with capsys.disabled():
-            print("\n  - ".join(results))
+            print(f"\n{benchman}")
+            benchman.print_results()
 
     # def test_bench_tree_mem(self, capsys):
     #     """ """